% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_time_series.R
\name{apply_plot_style}
\alias{apply_plot_style}
\title{Adds the default style layers to a ggplot object.}
\usage{
apply_plot_style(p)
}
\arguments{
\item{p}{A ggplot object.}

\item{color_palette}{A character string specifying the color palette.}

\item{include_scales}{Logical. Whether to include color/fill scales.}
}
\value{
A ggplot object with the style layers applied.
}
\description{
Adds the default style layers to a ggplot object.
}
