% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_planetscope_imagery.R
\name{build_cloud_cover_filter}
\alias{build_cloud_cover_filter}
\title{Build Cloud Cover Filter}
\usage{
build_cloud_cover_filter(cloud_lim = 0.1)
}
\arguments{
\item{cloud_lim}{Numeric. Maximum allowed cloud cover (0-1).}
}
\value{
A list representing the cloud cover filter.
}
\description{
Constructs a cloud cover filter.
}
