% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_planetscope_imagery.R
\name{build_date_range_filter}
\alias{build_date_range_filter}
\title{Build Date Range Filter}
\usage{
build_date_range_filter(date_start, date_end)
}
\arguments{
\item{date_start}{Character. Start date in "YYYY-MM-DD" format.}

\item{date_end}{Character. End date in "YYYY-MM-DD" format.}

\item{list_dates}{Optional vector of dates to refine the filter.}
}
\value{
A list representing the date range filter.
}
\description{
Constructs a date range filter for the search request.
}
