% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_planetscope_imagery.R
\name{build_geojson_from_bbox}
\alias{build_geojson_from_bbox}
\title{Build GeoJSON Geometry from Bounding Box}
\usage{
build_geojson_from_bbox(bbox)
}
\arguments{
\item{bbox}{A list with elements \code{xmin}, \code{ymin}, \code{xmax}, and \code{ymax}.}
}
\value{
A list representing the GeoJSON polygon.
}
\description{
Constructs a GeoJSON Polygon from a bounding box.
}
\examples{
bbox <- list(xmin = -77.05, ymin = 38.80, xmax = -76.90, ymax = 39.00)
geojson <- build_geojson_from_bbox(bbox)

}
