% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_planetscope_imagery.R
\name{build_gq_filter}
\alias{build_gq_filter}
\title{Build Ground Control and Quality Filter}
\usage{
build_gq_filter(ground_control = TRUE, quality = "standard")
}
\arguments{
\item{ground_control}{Logical. Whether ground control is required.}

\item{quality}{Character. Quality filter (default: "standard").}
}
\value{
A list representing the ground control and quality filter.
}
\description{
Constructs a filter for ground control and quality.
}
