% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_planetscope_imagery.R
\name{combine_filters}
\alias{combine_filters}
\title{Combine Search Filters}
\usage{
combine_filters(date_filter, cloud_filter, gq_filter, geometry_filter)
}
\arguments{
\item{date_filter}{List. Date range filter.}

\item{cloud_filter}{List. Cloud cover filter.}

\item{gq_filter}{List. Ground control and quality filter.}

\item{geometry_filter}{List. Geometry filter.}
}
\value{
A list representing the combined filter configuration.
}
\description{
Combines date range, cloud cover, ground control/quality, and geometry filters.
}
