% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_planetscope_imagery.R
\name{download_planetscope_imagery}
\alias{download_planetscope_imagery}
\title{Download a PlanetScope order}
\usage{
download_planetscope_imagery(
  order_id,
  exportfolder,
  api_key,
  overwrite = FALSE
)
}
\arguments{
\item{order_id}{Character. The PlanetScope order ID.}

\item{exportfolder}{Character. Directory in which to save downloaded files (created if needed).}

\item{api_key}{Character. Your Planet API key.}

\item{overwrite}{Logical. If \code{TRUE}, existing files will be overwritten (default: \code{FALSE}).}
}
\value{
Invisibly returns \code{NULL} and writes all imagery files to the specified \code{exportfolder}.
}
\description{
Downloads all files for a given PlanetScope order ID, saving them to the specified folder. The function will wait until the order is successfully processed by Planet API, partially processed, fail to be processed, or is cancelled.
}
\examples{
\dontrun{
download_planetscope_imagery(
  order_id = "abc123-order-id",
  exportfolder = "data/raw/SJER/SJER_202405_121_151",
  api_key = set_api_key(),
  overwrite = TRUE
)
}

}
