% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_planetscope_imagery.R
\name{download_planetscope_imagery_batch}
\alias{download_planetscope_imagery_batch}
\title{Download PlanetScope imagery for multiple sites and years}
\usage{
download_planetscope_imagery_batch(
  dir,
  v_site = NULL,
  v_year = 2017:(as.integer(format(Sys.Date(), "\%Y"))),
  v_month = 1:12,
  setting,
  num_cores = 12,
  overwrite = F
)
}
\arguments{
\item{dir}{Character. Base directory where the raw satellite data will be stored, expects a \verb{raw/} subdirectory.}

\item{v_site}{Character vector, optional. Site names to filter; if \code{NULL}, all sites are used.}

\item{v_year}{Numeric vector. Years for which to download data. Default is from 2017 to the current year.}

\item{v_month}{Integer vector. Months (1-12) to download (default: 1:12, i.e., all months).}

\item{setting}{List. Contains API settings including the API key (i.e., \code{setting$api_key}).}

\item{num_cores}{Integer. Number of parallel workers to use (default: 12).}

\item{overwrite}{Logical. If \code{TRUE}, existing files will be overwritten (default: \code{FALSE}).}
}
\value{
Invisibly returns \code{NULL} and saves downloaded imagery to the \verb{raw/} subdirectory of the specified \code{dir}.
}
\description{
Downloads all ordered PlanetScope imagery for specified sites and years. The function will wait until the order is successfully processed by Planet API, and then download the files.
}
\examples{
\dontrun{
download_planetscope_imagery_batch(
  dir = "alldata/PSdata/",
  v_site = c("HARV", "SJER"),
  v_year = 2025,
  v_month = 4:6,
  setting = setting,
  num_cores = 3,
  overwrite = FALSE
)
}

}
