% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_sample_data.R
\name{download_sample_data}
\alias{download_sample_data}
\title{Download sample data from GitHub}
\usage{
download_sample_data()
}
\value{
Path where the data was downloaded to.
}
\description{
This function clones the \code{sample-data} branch of the
\href{https://github.com/zhulabgroup/BatchPlanet}{BatchPlanet GitHub repository}
into the current working directory.
}
\details{
Requires that \code{git} is installed and available on your system PATH.
If the \code{sample-data} folder already exists, the function will not overwrite it.
}
