% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_planetscope_imagery.R
\name{extract_imagery_id}
\alias{extract_imagery_id}
\title{Fetch All Permissions Across Pagination}
\usage{
extract_imagery_id(initial_res, api_key, asset)
}
\arguments{
\item{initial_res}{List. The initial API response.}

\item{api_key}{Character. The API key.}

\item{asset}{Character. The asset type.}
}
\value{
A data frame containing permissions from all pages.
}
\description{
Retrieves all pages of search results and extracts download permissions.
}
