% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_planetscope_time_series.R
\name{retrieve_planetscope_time_series_sitegroup}
\alias{retrieve_planetscope_time_series_sitegroup}
\title{importFrom magrittr \%>\%}
\usage{
retrieve_planetscope_time_series_sitegroup(
  dir,
  df_coordinates,
  siteoi,
  groupoi,
  max_sample,
  num_cores
)
}
\description{
importFrom magrittr \%>\%
}
