% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_planetscope_time_series_batch.R
\name{retrieve_raster_data}
\alias{retrieve_raster_data}
\title{Process Reflectance or QA Raster Data}
\usage{
retrieve_raster_data(dir_site, sf_coordinates, type, num_cores)
}
\arguments{
\item{sf_coordinates}{sf object. Spatial points object.}

\item{type}{Character. Type of raster ("sr" for reflectance, "udm" for QA).}

\item{dir}{Character. Base directory for satellite data.}

\item{siteoi}{Character. Site identifier.}

\item{v_id}{Character vector. Unique IDs for the spatial points.}

\item{pattern}{Character. Regular expression pattern to match raster files.}
}
\value{
A data frame with extracted raster values.
}
\description{
Extracts data from raster files that match a given pattern (either reflectance or QA)
for a specified site.
}
