% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_whit.R
\name{util_fill_whit}
\alias{util_fill_whit}
\title{Whittaker smoothing wrapper}
\usage{
util_fill_whit(x, maxgap = Inf, lambda, minseg = 2)
}
\arguments{
\item{x}{signal to be smoothed: a vector}

\item{maxgap}{Maximum number of successive NAs to still perform imputation on. Default setting is to replace all NAs without restrictions.}

\item{lambda}{smoothing parameter: larger values lead to more smoothing}

\item{minseg}{Minimum length of segments to perform Whittaker smoothing on. Default setting is to 2.}
}
\value{
smoothed signal: a vector
}
\description{
This function performs Whittaker smoothing based on ptw:whit1. ptw::whit1 does not take NA in the time series. This function takes a time series with NA or placeholder value (e.g., -9999). Weights at placeholder should be set to 0. NA are skipped, while values at placeholders are interpolated.
}
\examples{
# Example:
data.frame(
  doy = 1:365, # simulate some data
  evi = sin((1:365) / 365 * pi)
) \%>\%
  filter(doy <= 20 | doy >= 40) \%>\% # make some short segments of data missing
  filter(doy <= 40 | doy >= 80) \%>\% # make some long segments of data missing
  complete(doy = seq(1, 365, by = 1)) \%>\%
  mutate(evi_sm = util_fill_whit(x = evi, maxgap = 14, lambda = 50, minseg = 2)) \%>\% # weighted whittaker smoothing allowing gaps \%>\%
  ggplot() +
  geom_line(aes(x = doy, y = evi_sm), col = "blue")
}
