% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{util_plot_style}
\alias{util_plot_style}
\title{Apply Default Plot Style}
\usage{
util_plot_style(p)
}
\arguments{
\item{p}{A ggplot object.}

\item{color_palette}{A character string specifying the color palette.}

\item{include_scales}{Logical. Whether to include color/fill scales.}
}
\value{
A ggplot object with the style layers applied.
}
\description{
Adds the default style layers to a ggplot object.
}
