% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_planetscope_imagery.R
\name{wait_for_order_success}
\alias{wait_for_order_success}
\title{Wait for Order to Succeed}
\usage{
wait_for_order_success(order_id, api_key)
}
\arguments{
\item{order_id}{Character. The order ID.}

\item{api_key}{Character. The API key.}
}
\value{
A list containing the order details when the state is 'success'.
}
\description{
Polls the remote sensing API until the order state is 'success', indicating that the order
is ready for download.
}
