% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_whit.R
\name{whitgap}
\alias{whitgap}
\title{Whittaker smoothing wrapper}
\usage{
whitgap(x, lambda, w)
}
\arguments{
\item{x}{signal to be smoothed: a vector}

\item{lambda}{smoothing parameter: larger values lead to more smoothing}

\item{w}{weights: a vector of same length as y. Default weights are equal to one}
}
\value{
smoothed signal: a vector
}
\description{
This function performs Whittaker smoothing based on ptw:whit1. ptw::whit1 does not take NA in the time series. This function takes a time series with NA or placeholder value (e.g., -9999). Weights at placeholder should be set to 0. NA are skipped, while values at placeholders are interpolated.
}
\examples{
# Example:
data.frame(
  doy = 1:365, # simulate some data
  evi = sin(doy)
) \%>\%
  filter(doy <= 20 | doy >= 40) \%>\% # make some data missing
  complete(doy = seq(1, 365, by = 1)) \%>\%
  mutate(evi_fill = zoo::na.fill(evi, fill = -9999, maxgap = 14)) \%>\% # fill short gaps with -9999 placeholder
  mutate(w = (evi_fill != -9999)) \%>\% # weight = 0 at long gaps, weight = 1 at short gaps
  mutate(evi_sm = whitgap(x = evi, lambda = 50, w = w)) \%>\% # weighted whittaker smoothing
  select(-evi_fill, -w)
}
