% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_planetscope_time_series.R
\name{clean_planetscope_time_series}
\alias{clean_planetscope_time_series}
\title{Clean a single PlanetScope time series}
\usage{
clean_planetscope_time_series(df_ts, calculate_evi)
}
\arguments{
\item{df_ts}{Data frame. Raw time series data for a single site/group.}

\item{calculate_evi}{Logical. If \code{TRUE}, computes Enhanced Vegetation Index (\code{evi}) after cleaning (default: \code{TRUE}).}
}
\value{
Data frame of cleaned time series, with EVI if requested.
}
\description{
Cleans a single time series data frame by removing low-quality data and optionally calculating EVI.
Low-quality data are defined as:
\itemize{
\item The sun elevation angle is less than 0 degrees (i.e., night time images).
\item The reflectance values for any band are less than 0.
\item The pixel was cloudy, had snow, ice, shadow, haze, or cloud.
\item The usable data mask had algorithmic confidence in classification less than 80\% for the pixel.
}
}
\examples{
\dontrun{
df_clean <- clean_planetscope_time_series(df_ts = df_ts_example, calculate_evi = TRUE)
}

}
