% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_seasonality.R
\name{determine_seasonality}
\alias{determine_seasonality}
\title{Evaluate Flatness of a Time Series Using AIC-based Model Selection}
\usage{
determine_seasonality(ts, doy = 1:length(ts), k = 50)
}
\arguments{
\item{ts}{Numeric vector. The time series signal to evaluate.}

\item{doy}{Numeric vector. The day-of-year indices corresponding to the time series.
Default is 1:length(ts).}

\item{k}{Numeric. Penalty parameter for the AIC calculation. Default is 50.}
}
\value{
Logical. TRUE if the straight-line model (with no breakpoints) is preferred;
FALSE otherwise. If the computation produces NA, it defaults to FALSE.
}
\description{
This function fits a simple linear regression model and segmented regression
models with 1 to 3 breakpoints to a given time series. It then compares their
Akaike Information Criterion (AIC) values (with a penalty parameter k) to assess
whether a straight-line (i.e., non-segmented) model is preferred. A return value
of TRUE indicates that the simple linear model is the best fit.
}
\examples{
\dontrun{
ts <- sin(seq(0, 2 * pi, length.out = 100)) + rnorm(100, sd = 0.1)
is_flat <- util_flat(ts)
}

}
