% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_planetscope_imagery.R
\name{download_planetscope_imagery}
\alias{download_planetscope_imagery}
\title{Download a Satellite Order}
\usage{
download_planetscope_imagery(
  order_id,
  exportfolder,
  api_key,
  overwrite = FALSE
)
}
\arguments{
\item{order_id}{Character. The order ID.}

\item{api_key}{Character. The API key.}

\item{order_name}{Character. The folder name in which to save the downloaded files.}

\item{order_num}{Numeric. Identifier for the order used for logging.}

\item{overwrite_opt}{Logical. Whether to overwrite existing files (default: FALSE).}
}
\value{
Invisibly returns NULL after downloading the files.
}
\description{
Retrieves the order details from the remote sensing API and downloads all associated items.
The function waits until the order is ready (state = 'success') before downloading.
}
