% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_planetscope_imagery.R
\name{order_planetscope_imagery}
\alias{order_planetscope_imagery}
\title{Place a PlanetScope order}
\usage{
order_planetscope_imagery(
  api_key,
  bbox,
  items,
  item_name,
  product_bundle,
  harmonized,
  order_name,
  mostrecent = 0
)
}
\arguments{
\item{api_key}{Character. Your Planet API key.}

\item{bbox}{An \code{sf}-style bounding box (a named numeric list with \verb{xmin,ymin,xmax,ymax}).}

\item{items}{Character vector. Planet image IDs to include in this order.}

\item{item_name}{Character. Item type (e.g. \code{"PSScene"}).}

\item{product_bundle}{Character. Product bundle (e.g. \code{"analytic_sr_udm2"}).}

\item{harmonized}{Logical. If \code{TRUE}, applies harmonization tool in the order.}

\item{order_name}{Character. A unique name for this order.}

\item{mostrecent}{Integer. If >0, only the most recent N images are ordered (default: 0 = all).}
}
\value{
Character. The ID of the order.
}
\description{
Submits an order to Planet API, with specified image IDs and tools.
}
\examples{
\dontrun{
order_id <- order_planetscope_imagery(
  api_key = set_api_key(),
  bbox = set_bbox(df_coordinates, "SJER"),
  items = ids,
  item_name = "PSScene",
  product_bundle = "analytic_sr_udm2",
  harmonized = TRUE,
  order_name = "SJER_20240501_20240531",
  mostrecent = 0
)
}

}
