% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_planetscope_imagery.R
\name{order_planetscope_imagery}
\alias{order_planetscope_imagery}
\title{Place a Satellite Order Request}
\usage{
order_planetscope_imagery(
  api_key,
  bbox,
  items,
  item_name,
  product_bundle,
  harmonized = FALSE,
  order_name = exportfolder,
  mostrecent = 0
)
}
\arguments{
\item{api_key}{Character. The API key.}

\item{bbox}{A list with bounding box values.}

\item{items}{Character vector. A vector of image IDs to order.}

\item{item_name}{Character. The imagery item type.}

\item{product_bundle}{Character. The product bundle to order.}

\item{harmonized}{Logical. Whether to include harmonization (default: FALSE).}

\item{order_name}{Character. A name for the order.}

\item{mostrecent}{Numeric. If greater than 0, only the specified number of most recent images are ordered.}
}
\value{
Character. The order ID if successful.
}
\description{
Submits an order request to the remote sensing API for a given set of image IDs.
}
