% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_planetscope_imagery_batch.R
\name{order_planetscope_imagery_siteyear}
\alias{order_planetscope_imagery_siteyear}
\title{Process Orders for a Single Site and Year}
\usage{
order_planetscope_imagery_siteyear(dir_site, siteoi, yearoi, bbox, setting)
}
\arguments{
\item{siteoi}{Character. Site name.}

\item{bbox}{Object. Bounding box for the site.}

\item{setting}{List. API settings.}

\item{path_sat_site}{Character. Path to the site's raw data directory.}

\item{year_download}{Numeric. Year to process.}
}
\value{
Invisibly returns NULL after saving the order summary.
}
\description{
For a given site and year, this function iterates through all 12 months, fetches available imagery
from the API based on the specified date range and bounding box, groups images into orders,
and saves an order summary.
}
