% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_planetscope_time_series.R
\name{retrieve_planetscope_time_series}
\alias{retrieve_planetscope_time_series}
\title{Process Satellite Data (Reflectance, QA, and Metadata)}
\usage{
retrieve_planetscope_time_series(dir_site, sf_coordinates, num_cores)
}
\arguments{
\item{sf_coordinates}{sf object. Spatial points corresponding to the site's coordinates.}

\item{dir}{Character. Base directory for satellite data.}

\item{siteoi}{Character. Site identifier.}

\item{v_id}{Character vector. Unique IDs for the spatial points.}
}
\value{
A data frame containing the combined satellite data.
}
\description{
Extracts and combines satellite data for a given site.
The function processes reflectance data, QA data, and metadata, merges these with point coordinate information,
applies scaling to the spectral bands, and returns a combined data frame.
}
