% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_planetscope_parameters.R
\name{set_api_key}
\alias{set_api_key}
\title{Set or change the Planet API key}
\usage{
set_api_key(change_key = F)
}
\arguments{
\item{change_key}{Logical. If \code{TRUE}, prompts for a new API key even if one already exists (default: \code{FALSE}).}
}
\value{
Invisibly returns the API key.
}
\description{
Prompts the user to enter a Planet API key and saves it in a hidden \code{.env} file in the working directory.
}
\examples{
\dontrun{
set_api_key() # Set the API key for the first time
set_api_key(change_key = TRUE) # Change the API key
}

}
