% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_planetscope_parameters.R
\name{set_planetscope_parameters}
\alias{set_planetscope_parameters}
\title{Set PlanetScope API Parameters}
\usage{
set_planetscope_parameters(
  api_key,
  item_name = "PSScene",
  asset = "ortho_analytic_4b_sr",
  product_bundle = "analytic_sr_udm2",
  cloud_lim = 1,
  harmonized = TRUE
)
}
\arguments{
\item{api_key}{Character. API key for authentication.}

\item{item_name}{Character. Name of the satellite data product (default: \code{"PSScene"}).}

\item{asset}{Character. Type of asset to retrieve (default: \code{"ortho_analytic_4b_sr"}).}

\item{product_bundle}{Character. Product bundle selection (default: \code{"analytic_sr_udm2"}).}

\item{cloud_lim}{Numeric. Cloud coverage limit (between 0 and 1, default: 1).}

\item{harmonized}{Logical. Indicates whether to use the Planset API tool to harmonize data (default: \code{TRUE}).}
}
\value{
A named list containing the PlanetScope API parameters.
}
\description{
Constructs a named list of parameters required for interacting with the PlanetScope API.
}
\examples{
\dontrun{
setting <- set_planetscope_parameters(
  api_key = set_api_key(),
  item_name = "PSScene",
  asset = "ortho_analytic_4b_sr",
  product_bundle = "analytic_sr_udm2",
  cloud_lim = 1,
  harmonized = TRUE
)
}

}
