% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_true_color_imagery.R
\name{visualize_coordinates}
\alias{visualize_coordinates}
\title{Visualize coordinates from a data frame}
\usage{
visualize_coordinates(df_coordinates)
}
\arguments{
\item{df_coordinates}{Data frame containing longitude and latitude columns.}
}
\value{
An interactive \code{plotly} object when supported, a static \code{ggplot} object otherwise.
}
\description{
Creates a scatter plot of coordinate data.
}
\examples{
\dontrun{
visualize_coordinates(df_coordinates)
}

}
