% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_true_color_imagery.R
\name{visualize_true_color_imagery}
\alias{visualize_true_color_imagery}
\title{Visualize Multi-Band Raster Data}
\usage{
visualize_true_color_imagery(file, df_coordinates = NULL, brightness = 5)
}
\arguments{
\item{brightness}{Numeric. Brightness multiplier (default: 1).}

\item{path}{Character. Path to the raster file.}

\item{crop_shape}{Optional. An \code{sf} or \code{SpatVector} object to crop the raster.}

\item{bands}{Character vector. Names of the RGB bands (default: c("red", "green", "blue")).}

\item{save_plot}{Logical. If TRUE, saves the plot.}

\item{plot_path}{Character. Directory in which to save the plot (default: "plots").}

\item{color_palette}{Character. The palette option for styling (default: "colorblind").}
}
\value{
A ggplot object displaying the RGB composite.
}
\description{
Creates an RGB composite visualization of a multi-band raster.
}
