% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whittaker_smoothing_filling.R
\name{whittaker_smoothing_filling}
\alias{whittaker_smoothing_filling}
\title{Whittaker Smoothing Wrapper}
\usage{
whittaker_smoothing_filling(x, lambda, maxgap = Inf, minseg = 2)
}
\arguments{
\item{x}{Numeric vector. The time series signal to be smoothed.}

\item{lambda}{Numeric. Smoothing parameter; larger values result in a smoother output.}

\item{maxgap}{Numeric. Maximum number of consecutive NAs to interpolate (default: Inf).}

\item{minseg}{Numeric. Minimum segment length for smoothing (default: 2).}
}
\value{
A numeric vector containing the smoothed signal.
}
\description{
Applies Whittaker smoothing to a time series. This function handles NA values and
placeholder values (-9999) by interpolating over short gaps.
}
\examples{
\dontrun{
df <- data.frame(doy = 1:365, evi = sin((1:365) / 365 * pi))
df <- df \%>\%
  filter(doy <= 20 | doy >= 40) \%>\%
  complete(doy = seq(1, 365, by = 1))
df <- df \%>\% mutate(evi_sm = util_fill_whit(x = evi, maxgap = 14, lambda = 50, minseg = 2))
ggplot(df, aes(x = doy, y = evi_sm)) +
  geom_line(col = "blue")
}
}
